Added support for lockfiles to make sure scripts are not running simultaneously (https://github.com/blaues0cke/PiSleepTalk/issues/80)

+++ bash/add-amplitude-to-filename.sh
@@ -14,52 +14,63 @@
 echo "Adding amplitude to filename"
 echo ""
 
-set file_counter=0
+script_name=`basename "$0"`
+lock_file_name="/var/lock/.${script_name}.exclusivelock"
 
-dir_list=$(ls ${audio_file_path_timestamp}/*.${default_audio_format} 2>/dev/null)
-for audio_file_path in $dir_list
-do
-	if [ -f $audio_file_path ]; then
-	 	echo "... processing file: $audio_file_path"
+if [ ! -d "${lock_file_name}" ]; then
+	mkdir "$lock_file_name"
 
-	 	audio_file_name=$(basename $audio_file_path)
+	set file_counter=0
 
-	 	# todo move to function
-	 	audio_timestamp=$(echo $audio_file_name | sed "s/\(\.${default_audio_format}\)//")
+	dir_list=$(ls ${audio_file_path_timestamp}/*.${default_audio_format} 2>/dev/null)
+	for audio_file_path in $dir_list
+	do
+		if [ -f $audio_file_path ]; then
+		 	echo "... processing file: $audio_file_path"
 
-	 	# todo: combine those two sox calls to one since both required amplitudes are retured with each call
+		 	audio_file_name=$(basename $audio_file_path)
 
-	 	maximum_amplitude=$(sox $audio_file_path -n stat 2>&1 | sed -n 's#^Maximum amplitude:[^0-9]*\([0-9.]*\)$#\1#p')
-	    midline_amplitude=$(sox $audio_file_path -n stat 2>&1 | sed -n 's#^Midline amplitude:[^0-9]*\([0-9.]*\)$#\1#p')
+		 	# todo move to function
+		 	audio_timestamp=$(echo $audio_file_name | sed "s/\(\.${default_audio_format}\)//")
 
-	 	echo "... maximum amplitude: $maximum_amplitude"
-	 	echo "... midline amplitude: $midline_amplitude"
+		 	# todo: combine those two sox calls to one since both required amplitudes are retured with each call
 
-		if [ -n "$maximum_amplitude" ] && [ -n "$midline_amplitude" ]; then
+		 	maximum_amplitude=$(sox $audio_file_path -n stat 2>&1 | sed -n 's#^Maximum amplitude:[^0-9]*\([0-9.]*\)$#\1#p')
+		    midline_amplitude=$(sox $audio_file_path -n stat 2>&1 | sed -n 's#^Midline amplitude:[^0-9]*\([0-9.]*\)$#\1#p')
 
-		 	new_audio_file_path="${audio_file_path_amplitude}/${audio_timestamp}-max-${maximum_amplitude}-mid-${midline_amplitude}.${default_audio_format}"
+		 	echo "... maximum amplitude: $maximum_amplitude"
+		 	echo "... midline amplitude: $midline_amplitude"
 
-		 	echo "... new file path: ${new_audio_file_path}"
+			if [ -n "$maximum_amplitude" ] && [ -n "$midline_amplitude" ]; then
 
-		 	mv $audio_file_path $new_audio_file_path
+			 	new_audio_file_path="${audio_file_path_amplitude}/${audio_timestamp}-max-${maximum_amplitude}-mid-${midline_amplitude}.${default_audio_format}"
 
-		else
+			 	echo "... new file path: ${new_audio_file_path}"
 
-			echo "... not able to get maximum and midline amplitude from file, file may be no audio file"
-			echo "... deleting ${audio_file_path}"
+			 	mv $audio_file_path $new_audio_file_path
 
-			rm $audio_file_path
+			else
+
+				echo "... not able to get maximum and midline amplitude from file, file may be no audio file"
+				echo "... deleting ${audio_file_path}"
+
+				rm $audio_file_path
+			fi
+
+			echo "... done"
+			echo ""
+
+			file_counter=$((file_counter + 1))
 		fi
+	done
 
-		echo "... done"
-		echo ""
-
-		file_counter=$((file_counter + 1))
+	if [ -n "$file_counter" ]; then
+	    echo "Done adding amplitude to filename, processed files: ${file_counter}"
+	else
+		echo "Done adding amplitude to filename, no files found";
 	fi
-done
-
-if [ -n "$file_counter" ]; then
-    echo "Done adding amplitude to filename, processed files: ${file_counter}"
+  	
+  	rmdir "$lock_file_name"
 else
-	echo "Done adding amplitude to filename, no files found";
+	echo "... done - existing lock file found"
 fi
